refactor: 축제 상세 클라이언트 컴포넌트 위치 이동
