+++ src/app/festival/_components/FestivalDetail.js
